ARROW-13244: [C++] Add facility to get current thread id as uint64
Followup to https://github.com/apache/arrow/pull/10632
